backend/requirements.txt
fastapi
uvicorn
python-multipart
rembg[u2net]
pillow
numpy
